⚡️  i actually think this hurts perf now ☉_☉ [b]
