+++ packages/tailwind-config/src/notion.js
@@ -1,3 +1,5 @@
+const radixColors = require('@radix-ui/colors')
+
 const notionColors = [
   // 'default',
   'gray',
@@ -20,11 +22,19 @@ const notionColors = [
   'red_background',
 ]
 
+/**
+ * @note(tailwind) you have lost your mind
+ */
+const globalNumbers = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
+const globalHack = []
+const globalDark = []
+
 const notion = {}
 notionColors.map((color) => {
-  // @note(radix-ui) map gray to slate
   let radixColor = color.split('_')[0]
-  if (radixColor === 'gray') radixColor = 'slate'
+  // // @note(radix-ui) map gray to slate
+  // if (radixColor === 'gray') radixColor = 'slate'
+
   if (color.includes('_background')) {
     notion['.notion-' + color] = {
       // @note(a11y) Radix-UI uses `4`, but Page Insights warns
@@ -34,7 +44,24 @@ notionColors.map((color) => {
     notion['.notion-' + color] = {
       color: `var(--${radixColor}11)`,
     }
+
+    // globalNumbers.map((n) => {
+    //   globalHack.push(
+    //     `--${radixColor}${n}: ${radixColors[radixColor][`${radixColor}${n}`]};`
+    //   )
+    //   globalDark.push(
+    //     `--${radixColor}${n}: ${
+    //       radixColors[`${radixColor}Dark`][`${radixColor}${n}`]
+    //     };`
+    //   )
+    // })
   }
 })
+
+// console.dir(`globalHack:`)
+// globalHack.map((g) => console.dir(g))
+// console.dir(`globalDark:`)
+// globalDark.map((g) => console.dir(g))
+// console.dir(`----`)
 
 module.exports = { notion, notionColors }
